Updated the JSON data to be an unnamed, verbose array.
There are now 5 functions available:

* "Named Array/Verbose": The codes array is named "codes" and each element is an object
with 2 fields: "code" and "text".

* "Named Array/Compressed": The codes array is named "codes" and each element is an
object with one field, with the code being the named field, and it's value that of the
text description.

* "Unnamed Array/Verbose": Same as "Named Array/Verbose" without naming the list of
enumerated codes.

* "Unnamed Array/Compressed": Same as "Named Array/Compressed", also lacking the name of
the enumerated codes.

* "All the Fields": All codes are compressed (see "Named Array/Compressed" for details)
as the individual fields of one object.

This generation of the JSON data now uses the "Unnamed Array/Verbose" method.